Consolidate code
