--- packages/react-dropdown-aria/styles/colours.ts
@@ -1,16 +1,0 @@
-const colours = {
-  greys: {
-    base: '#808080',
-    dark: '#7C7C7C',
-    darker: '#6e6d6d',
-    light: '#CECECE',
-    lighter: '#e0e0e0',
-    lightest: '#f5f5f5',
-  },
-  states: {
-    disabled: '#ededed',
-    focused: '#0080bc',
-  },
-};
-
-export default colours;
