+++ packages/react-dropdown-aria/utils/constants.ts
@@ -31,3 +31,18 @@ export const NAVIGATION_KEYS = [
 ];
 
 export const IdPrefix = 'react_dropdown_aria_';
+
+export const colours = {
+  greys: {
+    base: '#808080',
+    dark: '#7C7C7C',
+    darker: '#6e6d6d',
+    light: '#CECECE',
+    lighter: '#e0e0e0',
+    lightest: '#f5f5f5',
+  },
+  states: {
+    disabled: '#ededed',
+    focused: '#0080bc',
+  },
+};
